requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langsmith==0.3.10
multidict==6.1.0
numpy==1.26.4
ollama==0.4.7
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.38.6
tenacity==9.0.0
tiktoken==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
yarl==1.18.3
zstandard==0.23.0
